[WIP] Continue progress on HTMLBars integration
The highlight of this commit is that we now have an extremely
rudimentary view helper.

We also did a wide-ranging stream cleanup to allow streams to be
destroyed via refcounting (from below) rather than always destroying
from the top down.

This allows a lot of sharing in templates like:

```hbs
{{#if article.title}}
  <h1>{{article.title}}</h1>
{{/if}}

{{{article.body}}}
```

In this case, we could create only a single shared `article` stream,
a single shared `article.title` stream and a single `article.body`
stream. This takes advantage of the more robust notion of scope we got
from HTMLBars.