Refactor shared view and component logic
This commit moves most of the shared rendering logic for view and
components into the `ComponentNode` object. This object was already used
in the `component` hook, but it is now also used by the `view` keyword.

This commit also changes the `view` keyword to use the new lifecycle
style of keywords, which also means that it should now be idempotent.